+++ node/src/components/event_stream_server/tests.rs
@@ -200,6 +200,8 @@ struct TestFixture {
 impl TestFixture {
     /// Constructs a new `TestFixture` including `EVENT_COUNT` random events ready to be served.
     fn new(rng: &mut TestRng) -> Self {
+        const DISTINCT_EVENTS_COUNT: u32 = 7;
+
         let _ = logging::init();
         let storage_dir = tempfile::tempdir().unwrap();
         fs::create_dir_all(&storage_dir).unwrap();
@@ -207,7 +209,7 @@ impl TestFixture {
 
         let mut deploys = HashMap::new();
         let events = (0..EVENT_COUNT)
-            .map(|i| match i % 6 {
+            .map(|i| match i % DISTINCT_EVENTS_COUNT {
                 0 => SseData::random_block_added(rng),
                 1 => {
                     let (event, deploy) = SseData::random_deploy_accepted(rng);
@@ -215,9 +217,10 @@ impl TestFixture {
                     event
                 }
                 2 => SseData::random_deploy_processed(rng),
-                3 => SseData::random_fault(rng),
-                4 => SseData::random_step(rng),
-                5 => SseData::random_finality_signature(rng),
+                3 => SseData::random_deploy_expired(rng),
+                4 => SseData::random_fault(rng),
+                5 => SseData::random_step(rng),
+                6 => SseData::random_finality_signature(rng),
                 _ => unreachable!(),
             })
             .collect();